requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
defusedxml==0.4.1         # via python3-openid
django-allauth==0.30.0
django-debug-toolbar==1.6
Django==1.10.5            # via django-allauth, django-debug-toolbar
djangorestframework==3.5.3
environ==1.0
oauthlib==2.0.1           # via requests-oauthlib
python3-openid==3.0.10    # via django-allauth
requests-oauthlib==0.7.0  # via django-allauth
requests==2.12.4          # via django-allauth, requests-oauthlib
sqlparse==0.2.2           # via django-debug-toolbar
